completely redid file saving. instead of saving it as a plist or json file, it will now save as a sceneproj file. right now, it's structured like this: myfile.sceneproj
--project.plist OR project.json (will change according to save settings)
--resources
----resource1.png
----resource2.png

this has multiple benefits. now, when adding the project to your game, you don't need to worry about the resources, since they're automatically copied into the file. additionally, if the file you imported into your scene is deleted or changed, the scene project file will retain it for you so that when you load up your project, it is still intact.